Move `{Node,Edge}Type` to src/analysis/types.js
Test plan: `yarn test`

Part of ongoing work on #704
